mv mkConfig into catalog app
